added backend and ml code
