+++ Frontend/regressive-visualization-frontend/e2e/protractor.conf.js
