+++ Frontend/regressive-visualization-frontend/e2e/tsconfig.json
