+++ Frontend/regressive-visualization-frontend/karma.conf.js
