+++ Frontend/regressive-visualization-frontend/package.json
@@ -22,7 +22,9 @@
     "@fortawesome/angular-fontawesome": "^0.6.1",
     "@fortawesome/fontawesome-svg-core": "^1.2.28",
     "@fortawesome/free-solid-svg-icons": "^5.13.0",
+    "@ng-bootstrap/ng-bootstrap": "^5.3.1",
     "angular-particle": "^1.0.4",
+    "bootstrap": "^4.4.1",
     "chart.js": "^2.9.3",
     "rxjs": "~6.4.0",
     "tslib": "^1.10.0",
