+++ Frontend/regressive-visualization-frontend/src/assets/logo.png
