+++ Frontend/regressive-visualization-frontend/src/styles.css
